Move secretproviderclass.go to support folder
This commit moves the secretproviderclass.go and its unit test file to
the support directory. The function,
ReconcileManagedAzureSecretProviderClass, within this file is called by
both the HO and the CPO.
